Factored out topics/complex_numbers constructs to the appropriate places with mobject/, animation/ and scene/
